Make use of pydantic for configs.
This comes with some necessary changes to support the change in
configuration handling, such as allowing the user to specify which
class they are using for the Client property in modules.

This means that the Config class can be subclassed to support
custom attributes (assuming a subclass of discord.Client or
eris.Core is used).